refactor pool code: move math to markets.py, quotes to queries.py, remove wrappers
- Move pool utilities (sort_pair, compute_pool_output, compute_pool_input_for_target_price,
  get_pool_fee_bps, pool_has_liquidity, isqrt) from messages/pool.py to ledger/markets.py,
  matching how issuances.price() lives alongside asset data access
- Slim messages/pool.py to matching-only (try_pool_fill, execute_pool_match,
  fill_resting_orders, match_resting_orders_against_pool)
- Remove 6 passthrough wrappers (get_pool, insert_pool, update_pool, etc.) that
  no other message file uses — callers now use ledger.markets.* directly
- Move quote functions from compose.py to queries.py using state_db (consolidated
  tables), eliminating the ledger import that compose.py never needed
- Replace raw SQL in quotes with select_rows/select_row helpers matching codebase
  conventions (orders_info view with give_price sort, assets_info for LP supply)
- Fix quote fee_amount to only count pool input, not book order fills
- Update routes.py to point quote endpoints to queries module
- Update all committed tests to use ledger.markets.* for data access
